Move it out so it can be tested.
